A new function in Helpful Functions
This commit contains a new function that allows you to convert milliseconds into a readable format. Please note that the code is rather messy, but it works quite well. Also note that the calculation of days is set to 30, since it's not always possible to know how many exact days a month will have.