Simple Password Maker Disclaimer
Adds a warning to inform everyone that this script should not be used as a password generator for protecting confidential information. Please use legit password managers that know what they are doing. For instance, you could use this for a TeamTalk or game password, but don't go using it on your social media or financial accounts.