05/29/2025
Adds a key_hold script for Pygame. Since it goes hand in hand with the pygame timer, it is under Python/utils/timers. A test script called "key_hold test.py" has been included, which will play the Windows ding sound when pressing and/or holding the Space key down.